% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{add_row_hash}
\alias{add_row_hash}
\title{Add hash for each row}
\usage{
add_row_hash(df, primary_keys)
}
\arguments{
\item{df}{a data.frame}

\item{primary_keys}{the column anmes of the primary key}
}
\description{
It sorts the column names, it hash every row and add the column.
}
\examples{
df <- data.frame(
  id = c(1, 2, 3),
  name = c("AAAAA", "BBBB", "CCC")
)
add_row_hash(df, id)
}
\keyword{wrangling}
