% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{addin_generate_documentation}
\alias{addin_generate_documentation}
\title{Adding to generate documentation}
\usage{
addin_generate_documentation(
  model = "qwen2.5-coder:3b",
  context = rstudioapi::getActiveDocumentContext()
)
}
\arguments{
\item{model}{A single string with the ollama model to use.}

\item{context}{the IDE context. Defaults to rstudioapi::getActiveDocumentContext}
}
\value{
Nothing.
}
\description{
It changes the selected code for the the generated documentation using
the configured ollama model.
}
\keyword{packages}
