% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{aside}
\alias{aside}
\title{Multiple aside functions with base R pipe}
\usage{
aside(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{functions to run aside}
}
\description{
Multiple aside functions with base R pipe
}
\examples{
n_try <- 1
rnorm(200) |>
  matrix(ncol = 2) |>
  aside(
    print("Matrix prepared"),
    print(n_try)
  ) |>
  colSums()

}
\keyword{pipe}
