% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_split_prop}
\alias{calc_split_prop}
\title{Calculate split proportion}
\source{
\url{https://stats.stackexchange.com/a/305063/7387}
}
\usage{
calc_split_prop(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
From a data frame, it returns the split proportion.
}
\details{
The calc_validation_size function returns the optimal split
proportion according to the number of rows for your validation set.
}
\examples{
calc_split_prop(data.frame(row = 1:891))

}
\keyword{tidymodels}
