% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_split_size}
\alias{calc_split_size}
\title{Calculate split size}
\source{
\url{https://stats.stackexchange.com/a/304996/7387}
}
\usage{
calc_split_size(std_err = 0.001)
}
\arguments{
\item{std_err}{The desired std_err numeric (default 0.001)}
}
\description{
From binary classification problems, with the desired std_err it returns the
minimal assesment/validation set size.
}
\details{
The calc_validation_size function returns the minimal validation size for expected probabilities and desired error.
s
}
\examples{
calc_split_size()
calc_split_size(std_err = 0.02)

}
\keyword{tidymodels}
