% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_validation_size}
\alias{calc_validation_size}
\title{Calculate validation size}
\source{
\url{https://stats.stackexchange.com/a/304996/7387}
}
\usage{
calc_validation_size(expected_p, desired_std_err)
}
\arguments{
\item{expected_p}{An object}

\item{desired_std_err}{An expresion}
}
\description{
From and expected p and desired std_err, it returns the
minimal validation size for your assesment sets or validation sets.
}
\details{
The calc_validation_size function returns the minimal validation size for expected probabilities and desired error.
}
\examples{
calc_validation_size(expected_p = 0.8, desired_std_err = 0.02)

}
\seealso{
\url{https://github.com/jrosell/jrrosell/blob/main/R/calc.R}
}
