% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{chars}
\alias{chars}
\title{Create a vector of characters from a string}
\usage{
chars(x, ...)
}
\arguments{
\item{x}{a vector of characters of length 1.}

\item{...}{unused}
}
\value{
a vector of characters
}
\description{
Create a vector of characters from a string
}
\details{
\code{chars} expects a single string as input. To create a list of these,
consider \code{lapply(strings, chars)}.
}
\seealso{
\url{https://github.com/jonocarroll/charcuterie}
}
\keyword{types}
