% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{check_installed_gihub}
\alias{check_installed_gihub}
\title{Check if the last github version is installed}
\usage{
check_installed_gihub(repo)
}
\arguments{
\item{repo}{a github repo/package. Ex: check_installed_gihub("tidyverse/dplyr")}
}
\description{
Get the version from the DESCRIPTION file of the master branch in the github package.
}
\keyword{packages}
