% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{check_installed_github}
\alias{check_installed_github}
\title{Check if the last github version is installed}
\usage{
check_installed_github(repo)
}
\arguments{
\item{repo}{a github repo/package. Ex: check_installed_github("tidyverse/dplyr")}
}
\description{
Check if the last main github version is installed.
}
\examples{
if (FALSE) {
  check_installed_github("jrosell/jrrosell")
}
}
\keyword{packages}
