% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{count_duplicated_rows}
\alias{count_duplicated_rows}
\title{Count the number of duplicated rows}
\usage{
count_duplicated_rows(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Count the number of duplicated rows
}
\examples{
count_duplicated_rows(data.frame(a = c(1, 2, 3), b = c(3, 4, 5)))
count_duplicated_rows(data.frame(a = c(1, 2, 3), b = c(1, 4, 5)))
}
\keyword{wrangling}
