% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{detect_cores}
\alias{detect_cores}
\title{Detect cores that could be used}
\usage{
detect_cores(max = 10, min = 2)
}
\arguments{
\item{max}{An integer with the max desired cores (default 10)}

\item{min}{An integer with the min desired cores (default 2)}
}
\description{
Select cores in max/min of the available cores.
}
\details{
The detect_cores function uses parallelly
package. It returns the desired max cores if available or it fails if not
min cores are available.
}
\examples{
cores <- detect_cores(max = 5, min = 1)
print(cores)
if (FALSE) {
  library(jrrosell)
  library(future)
  plan(multisession, workers = detect_cores(max = 10, min = 2))
  plan(sequential)
}
}
\keyword{parallel}
