% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{.get_recipe_mode}
\alias{.get_recipe_mode}
\title{Help infer things about the mode and the model to fit}
\usage{
.get_recipe_mode(rec)
}
\description{
Help infer things about the mode and the model to fit
}
\keyword{internal}
