% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{count_duplicated_rows}
\alias{count_duplicated_rows}
\title{Find the number of duplicated rows}
\usage{
count_duplicated_rows(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Find the number of duplicated rows
}
\keyword{processing}
