% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{end_pipe}
\alias{end_pipe}
\title{End a pipe}
\usage{
end_pipe(x, expr)
}
\arguments{
\item{x}{An object}

\item{expr}{An expresion}
}
\description{
Print and expression and return invisible NULL at the end of a pipe.
}
\examples{
library(ggplot2)
rnorm(200) |>
  matrix(ncol = 2) |>
  as.data.frame() |>
  end_pipe({
    ggplot(x, aes(V1, V2)) +
      geom_point()
  })

}
\keyword{pipe}
