% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{generate_documentation}
\alias{generate_documentation}
\title{Generate documentation}
\usage{
generate_documentation(x, model = "qwen2.5-coder:3b")
}
\arguments{
\item{x}{A single string.}

\item{model}{A single string with the ollama model to use.}
}
\value{
Generated documentation as a character string.
}
\description{
It returns the genereated documentation from the selected model
}
\keyword{packages}
