% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{get_sentiments_by_language}
\alias{get_sentiments_by_language}
\title{Get a sentiments by language}
\usage{
get_sentiments_by_language(language = "en", lexicon = "chen_skiena")
}
\arguments{
\item{language}{two letters language code.}

\item{lexicon}{default and only valid value "chen_skiena"}
}
\value{
A tibble with word and sentiment columns
}
\description{
The multilingual sentiment lexicon was obtained from here on 2024-12-18 https://aclanthology.org/P14-2063/
}
\details{
The files were generated this way:
chen_skiena_lexicon <-
bind_rows(
here::here("P14-2063.Datasets", "readable_neg_words_list.txt") |>
read_delim(delim = " ", col_names = c("word", "lang")) |>
mutate(sentiment = factor("negative", levels = c("negative", "positive"))),
here::here("P14-2063.Datasets", "readable_pos_words_list.txt") |>
read_delim(delim = " ", col_names = c("word", "lang")) |>
mutate(sentiment = factor("positive", levels = c("negative", "positive")))
)
chen_skiena_lexicon |>
write_fst(
here::here("inst", "extdata", "chen_skiena_lexicon.fst"),
compress = 100
)
top_languages <- rlang::chr(
ca = 'catalan',
zh = 'chinese_simplified',
da = 'danish',
nl = 'dutch',
en = 'english',
eo = 'esperanto',
fi = 'finnish',
fr = 'french',
de = 'german',
el = 'greek',
hu = 'hungarian',
it = 'italian',
la = 'latin',
pt = 'portuguese',
es = 'spanish',
sv = 'swedish'
)
nrc_lexicon <- read_delim("NRC-Emotion-Lexicon-ForVariousLanguages.txt", delim = "\\t") |>
janitor::clean_names() |>
pivot_longer(cols = anger:trust, names_to = "sentiment") |>
rename(english = english_word) |>
pivot_longer(cols = -c(sentiment, value), names_to = "language", values_to = "word") |>
filter(sentiment \%in\% c("positive", "negative")) |>
filter(language \%in\% top_languages) |>
transmute(
sentiment = factor(sentiment, c("negative", "positive")),
language = factor(language, unique(language)),
word = factor(word, unique(word)),
) |>
glimpse()
nrc_lexicon |>
write_fst(
here::here("nrc_lexicon.fst"),
compress = 100
)
}
\examples{
get_sentiments_by_language("ca")
}
\seealso{
\url{https://juliasilge.github.io/tidytext/reference/get_sentiments.html}
}
\keyword{text}
