% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{glimpse_variable}
\alias{glimpse_variable}
\title{Glimpse a variable}
\usage{
glimpse_variable(df, variable)
}
\arguments{
\item{df}{a data.frame}

\item{variable}{the variable to use.}
}
\description{
It returns the ordered counts of the variable in the data.frame.
}
\keyword{processing}
