% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{glimpses}
\alias{glimpses}
\title{Glimpse multiple datasets}
\usage{
glimpses(...)
}
\arguments{
\item{...}{Multiple data.frame}
}
\description{
Glimpse multiple datasets
}
\examples{
df1 <- data.frame(a = c(1, 2))
df2 <- data.frame(b = c(3, 4))
glimpses(df1, df2)
}
\keyword{wrangling}
