% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrrosell-package.R
\docType{package}
\name{jrrosell-package}
\alias{jrrosell}
\alias{jrrosell-package}
\title{jrrosell: Personal R package for Jordi Rosell}
\description{
Useful functions for personal usage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrosell/jrrosell}
  \item Report bugs at \url{https://github.com/jrosell/jrrosell/issues}
}

}
\author{
\strong{Maintainer}: Jordi Rosell \email{jroselln@gmail.com} (\href{https://orcid.org/0000-0002-4349-1458}{ORCID})

}
\keyword{internal}
