% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{kgl_competitions_data_download}
\alias{kgl_competitions_data_download}
\title{Download files from kaggle competition}
\source{
\url{https://github.com/mkearney/kaggler/issues/6}
}
\usage{
kgl_competitions_data_download(competition_id, file_name, path = ".")
}
\arguments{
\item{competition_id}{Number Id of the Kaggle competition.}

\item{file_name}{String with the file you want to download.}

\item{path}{String with the folder where you want the file.}
}
\description{
Download files from kaggle competition
}
\seealso{
\url{https://github.com/jrosell/jrrosell/blob/main/R/kaggle.R}
}
