% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{last_fit_metrics}
\alias{last_fit_metrics}
\title{Do the last fit and get the metrics}
\usage{
last_fit_metrics(res, split, metric)
}
\arguments{
\item{res}{Tune results}

\item{split}{The initial split object}

\item{metric}{What metric to use to select the best workflow}
}
\description{
Do the last fit and get the metrics
}
\keyword{tidymodels}
