% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{name_unnamed_chunks}
\alias{name_unnamed_chunks}
\title{Name unnamed chunks in .Rmd or .qmd files
Use with caution.}
\usage{
name_unnamed_chunks(file_path)
}
\arguments{
\item{file_path}{the file name}
}
\description{
Name unnamed chunks in .Rmd or .qmd files
Use with caution.
}
\keyword{processing}
