% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{normalize_text}
\alias{normalize_text}
\title{Normalize text}
\usage{
normalize_text(text, remove_digits = TRUE, remove_accents = TRUE)
}
\arguments{
\item{text}{A character vector or object that can be coerced to a character string. Represents the input text to be cleaned.}

\item{remove_digits}{= TRUE}

\item{remove_accents}{= TRUE}
}
\value{
A normalized character vector
}
\description{
This function processes a given text string by converting it to lowercase, removing numbers,
non-alphanumeric characters, extra whitespace.
It also transliterates text to ASCII, splits words, and reconstructs a clean text string suitable for analysis.
}
\examples{
# Example usage:
normalize_text("¡Hola! Esto es una prueba 123.")

}
\keyword{text}
