% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{normalize_vec}
\alias{normalize_vec}
\title{Center and scale double vectors}
\usage{
normalize_vec(...)
}
\arguments{
\item{...}{a double vector or multiple double vectors}
}
\description{
Center and scale double vectors
}
\examples{
normalize_vec(1, 2, 3, )
}
\keyword{tidymodels}
