% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{notify_finished}
\alias{notify_finished}
\title{Make a sound and send an email when a process finished}
\usage{
notify_finished(name, body = "", ..., sound = 1, tictoc_result = NULL)
}
\arguments{
\item{name}{The process name (Required)}

\item{body}{The contents of the email (Default "")}

\item{...}{Additional arguments to pass to the template function. If you're using the default template, you can use font_family to control the base font, and content_width to control the width of the main content; see blastula_template(). By default, the content_width is set to ⁠1000px⁠. Using widths less than ⁠600px⁠ is generally not advised but, if necessary, be sure to test such HTML emails with a wide range of email clients before sending to the intended recipients. The Outlook mail client (Windows, Desktop) does not respect content_width.}

\item{sound}{The sound for beepr::beep call (Default 1)}

\item{tictoc_result}{the result from tictoc::toc (Default NULL)}
}
\description{
The notify_finished make a sound using beepr::beep, compose and email and send it returing the blastula::smtp_send call results.
}
\details{
The following environment variables should be set:
\itemize{
\item MY_SMTP_USER from
\item MY_SMTP_RECIPIENT to
\item MY_SMTP_PASSWORD service password (for gmail you can use https://myaccount.google.com/apppasswords)
\item MY_SMTP_PROVIDER blastula provider (gmail if not set)
}
}
\examples{
if (exists("not_run")) {
  tictoc::tic()
  Sys.sleep(1)
  jrrosell::notify_finished("job", "Well done", sound = "fanfare", tictoc_result = tictoc::toc())
}

}
\keyword{processing}
