% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{package_github_name}
\alias{package_github_name}
\title{Github name of the package}
\usage{
package_github_name(x, file_lines = NULL)
}
\arguments{
\item{x}{a single repo/package to check Ex: package_github_name("tidyverse/dplyr")}

\item{file_lines}{(default = NULL, internal)}
}
\description{
Get the name of the package from the DESCRIPTION file of the master branch in the github repo
}
\seealso{
https://github.com/jrosell/jrrosell/blob/main/R/packages.R
}
\keyword{packages}
