% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{plot_histograms}
\alias{plot_histograms}
\title{Plot histograms for double columns}
\usage{
plot_histograms(df, ...)
}
\arguments{
\item{df}{a data.frame}

\item{...}{optional parameters to geom_histogram}
}
\description{
Plot histograms for double columns
}
\keyword{processing}
