% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{plot_missing}
\alias{plot_missing}
\title{Plot missing values}
\usage{
plot_missing(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Plot missing values
}
\examples{
plot_missing(data.frame(a = c(1, NA), b = c(NA, 4)))
}
\keyword{wrangling}
