% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot a variable}
\usage{
plot_variable(df, variable)
}
\arguments{
\item{df}{a data.frame}

\item{variable}{the variable to use.}
}
\description{
It returns a bar or a histogram of the variable
}
\keyword{processing}
