% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{prep_juice}
\alias{prep_juice}
\title{Prep, juice and glimpse a recipe or workflow}
\source{
\url{https://recipes.tidymodels.org/reference/update.step.html}
}
\usage{
prep_juice(object)
}
\arguments{
\item{object}{A recipe or a workflow object with a recipe}
}
\description{
Prep, juice and glimpse a recipe or workflow
}
\keyword{tidymodels}
