% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{prep_juice_cols}
\alias{prep_juice_cols}
\title{Prep, juice and get cols from a recipe or workflow}
\usage{
prep_juice_cols(object)
}
\arguments{
\item{object}{A recipe or a workflow object with a recipe}
}
\description{
Prep, juice and get cols from a recipe or workflow
}
\examples{
recipes::recipe(spray ~ ., data = InsectSprays) |>
   prep_juice_cols()

}
\seealso{
https://github.com/jrosell/jrrosell/blob/main/R/tidymodels.R
}
