% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{prep_juice_ncol}
\alias{prep_juice_ncol}
\title{Prep, juice and get cols from a recipe or workflow}
\usage{
prep_juice_ncol(object)
}
\arguments{
\item{object}{A recipe or a workflow object with a recipe}
}
\description{
Prep, juice and get cols from a recipe or workflow
}
\examples{
recipes::recipe(spray ~ ., data = InsectSprays) |>
  prep_juice_ncol()
}
\keyword{tidymodels}
