% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{prepare_docs}
\alias{prepare_docs}
\title{Prepare docs for Analysis}
\usage{
prepare_docs(df, ...)
}
\arguments{
\item{df}{data frame with and id and text columns.}

\item{...}{paramters passed to \code{"prepare_tokens"}}
}
\value{
A df with a list of tokens and character vector prepared_text columns for documents at column id and text at column "text"
}
\description{
Prepare docs for Analysis
}
\examples{
# Example usage:
prepare_docs(data.frame(id = 1, text = "¡Hola! Esto es una prueba 123."))

}
\keyword{text}
