% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{prepare_text}
\alias{prepare_text}
\title{Prepare Text for Analysis}
\usage{
prepare_text(text, stopwords = NULL)
}
\arguments{
\item{text}{A character vector or object that can be coerced to a character string. Represents the input text to be cleaned.}

\item{stopwords}{A character vector specifying stopwords removal. Defaults to \code{"spanish"} stopwords from the tm:stopwords package.}
}
\value{
A cleaned character string, with stopwords removed and text formatted for analysis.
}
\description{
This function processes a given text string by converting it to lowercase, removing numbers,
non-alphanumeric characters, extra whitespace, and stopwords based on a specified language.
It also transliterates text to ASCII, splits words, and reconstructs a clean text string suitable for analysis.
}
\examples{
# Example usage:
prepare_text("¡Hola! Esto es una prueba 123.")

}
\keyword{text}
