% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_xlsx}
\alias{read_xlsx}
\title{Read a xlsx file into a a list of tibbles}
\usage{
read_xlsx(xlsxFile, ..., sheet = 1, startRow = 1)
}
\arguments{
\item{xlsxFile}{The name of the file.}

\item{...}{Other parameters to openxls::read.xlsx function}

\item{sheet}{The name or index of the sheet (default 1)}

\item{startRow}{The number of the starting reading row (default 1)}
}
\description{
It's useful for saving multiple data to a multiple sheets of a single Excel/Openoffice file.
}
\details{
The write_xlsx it's a wroapper for \code{openxls::write.xlsx}.
}
\examples{
l <- list("IRIS" = iris, "MTCATS" = mtcars, matrix(runif(1000), ncol = 5))
write_xlsx(l, "writeList1.xlsx", colWidths = c(NA, "auto", "auto"))
read_xlsx("writeList1.xlsx")
file.remove("writeList1.xlsx")

}
\seealso{
\url{https://github.com/jrosell/jrrosell/blob/main/R/read.R}
}
