% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{rebuild_package_and_check}
\alias{rebuild_package_and_check}
\title{Internal helper function for package development}
\usage{
rebuild_package_and_check()
}
\description{
Internal helper function for package development
}
\examples{
if (FALSE) {
  devtools::load_all(); rebuild_package_and_check()
}

}
\keyword{internal}
