% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{remove_stopwords}
\alias{remove_stopwords}
\title{Remove stopwords}
\usage{
remove_stopwords(text, stopwords = NULL, lang = "spanish")
}
\arguments{
\item{text}{A character vector or object that can be coerced to a character string. Represents the input text to be cleaned.}

\item{stopwords}{A character vector specifying stopwords removal. Defaults tm:stopwords package.}

\item{lang}{defaults to \code{"spanish"}}
}
\value{
A character vector without stopwords
}
\description{
This function processes character vectors and remove the specified stop words
or the stoop words of the langauge from the tm package
}
\examples{
# Example usage:
normalize_text("¡Hola! Esto es una prueba 123.") |>
  tokenize_text() |>
  remove_stopwords()

}
\keyword{text}
