% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{request_max_safe_cores_from_rss}
\alias{request_max_safe_cores_from_rss}
\title{Request the maximum safe number of cores}
\usage{
request_max_safe_cores_from_rss(
  estimated_max_rss,
  memory_usage = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{estimated_max_rss}{butes of maximum rss it will used (You can get it from syrup package)}

\item{memory_usage}{the proportion of the system memory that will be used (0.8)}

\item{verbose}{to debug (TRUE)}
}
\description{
When parallizing within resamples, required memory can crash the system.
}
\details{
The detect_cores function uses parallelly
package. It returns the desired max cores if available or it fails if not
min cores are available (excluding system reserved cores).
}
\keyword{parallel}
