% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{resp_body_yyjson}
\alias{resp_body_yyjson}
\title{Extract body from httr2 response using yyjsonr}
\usage{
resp_body_yyjson(resp, check_type = TRUE, simplifyVector = FALSE, ...)
}
\arguments{
\item{resp}{A httr2::response object, created by httr2::req_perform().}

\item{check_type}{Should the type actually be checked? Provided as a
convenience for when using this function inside \verb{resp_body_*} helpers.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{...}{Other parameters}
}
\description{
Extract body from httr2 response using yyjsonr
}
\keyword{read}
