% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{sanitize_title_with_dashes}
\alias{sanitize_title_with_dashes}
\title{Sanitize title with dashes}
\usage{
sanitize_title_with_dashes(title)
}
\arguments{
\item{title}{the title}
}
\description{
It generates slugs URLs as WordPress does
}
\examples{
sanitize_title_with_dashes("Hello world")
}
\keyword{wrangling}
