% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{select_constant_columns}
\alias{select_constant_columns}
\title{Select constant columns from a data.frame}
\usage{
select_constant_columns(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Select constant columns from a data.frame
}
\keyword{processing}
