% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{slugify}
\alias{slugify}
\title{Slugify character vectors}
\usage{
slugify(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
It generates slug URLs handling ASCII normalization
}
\examples{
sanitize_title_with_dashes("Hello world")
}
\keyword{wrangling}
