% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spain_ccaas}
\alias{spain_ccaas}
\title{spain_ccaas}
\format{
\subsection{\code{spain_ccaas}}{

A sf object with 19 rows and 4 columns:
\describe{
\item{OBJECTID}{}
\item{codigo}{}
\item{nombre}{}
\item{geometry}{}
}
}
}
\source{
\url{https://github.com/koldLight/curso-r-dataviz/blob/master/dat/spain_ccaas.geojson}
}
\usage{
spain_ccaas
}
\description{
spain_ccaas
}
\examples{
library(sf)
data(spain_ccaas)
head(spain_ccaas)
}
\seealso{
\url{https://github.com/jrosell/jrrosell/blob/main/data/spain_ccaas.rda}
}
\keyword{datasets}
