% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spain_provinces}
\alias{spain_provinces}
\title{spain_provinces}
\format{
\subsection{\code{spain_provinces}}{

A sf object with 60 rows and 4 columns:
\describe{
\item{OBJECTID}{}
\item{codigo}{}
\item{nombre}{}
\item{geometry}{}
}
}
}
\source{
\url{https://github.com/koldLight/curso-r-dataviz/blob/master/dat/spain_provinces.geojson}
}
\usage{
spain_provinces
}
\description{
spain_provinces
}
\examples{
library(sf)
data(spain_provinces)
head(spain_provinces)
}
\seealso{
\url{https://github.com/jrosell/jrrosell/blob/main/data/spain_provinces.rda}
}
\keyword{datasets}
