% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{startParallel}
\alias{startParallel}
\title{Start parallel processing}
\usage{
startParallel(msg = NULL, max = 10, min = 1)
}
\arguments{
\item{msg}{A character vector with the tictoc timer message.}

\item{max}{An integer with the max desired cores.}

\item{min}{An integer with the min desired cores.}
}
\description{
Start timer and parallel processing in max/min of the available cores.
}
\details{
The startParallel function uses tictoc, parallelly, parallel
packages. It returns a parallel cluster to be passed as a stopParallel
argument
}
\examples{
cl <- startParallel("parallel processing", max = 3, min = 1)
print("parallel processing here")
stopParallel(cl)

}
\keyword{parallel}
