% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{stopParallel}
\alias{stopParallel}
\title{Stop parallel processing}
\usage{
stopParallel(cl)
}
\arguments{
\item{cl}{A cluster object returned by startParallel or parallel::makeCluster, parallel::makePSOCKcluster, parallel::makeForkCluster}
}
\description{
Stop started timer and parallel processing.
}
\details{
The stopParallel function uses tictoc and  parallel
packages. It returns the result of parallel::stopCluster(cl) method.
}
\examples{
cl <- startParallel()
print("parallel processing here")
stopParallel(cl)

}
\seealso{
https://github.com/jrosell/jrrosell/blob/main/R/parallel.R
}
\keyword{parallel}
