% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{sum_missing}
\alias{sum_missing}
\title{Sum the missing values from a data.frame}
\usage{
sum_missing(...)
}
\arguments{
\item{...}{one or multiple data.frame}
}
\description{
Sum the missing values from a data.frame
}
\examples{
sum_missing(data.frame(a = c(1, 2), b = c(3, 4)))
sum_missing(data.frame(a = c(1, NA), b = c(3, 4)))
sum_missing(data.frame(a = c(1, NA), b = c(NA, 4)))
sum_missing(data.frame(a = c(NA, NA), b = c(NA, NA)))
}
\keyword{wrangling}
