% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{tee}
\alias{tee}
\title{Tee pipe}
\source{
\url{https://mastodon.social/@multimeric@genomic.social/109555362766969210}
}
\usage{
tee(x, expr)
}
\arguments{
\item{x}{An object}

\item{expr}{An expresion}
}
\description{
Pipe a value forward into a function- or call expression and return the
original value instead of the result. This is useful when an expression
is used for its side-effect, say plotting or printing.
}
\details{
The tee pipe works like \code{|>}, except the
return value is \code{x} itself, and not the result of \code{expr} call.
}
\section{Thanks}{


I want to give credit to Michael Milton and Matthew Kay for the idea and the code.
}

\examples{
rnorm(200) |>
    matrix(ncol = 2) |>
    tee(plot) |> # plot usually does not return anything.
    colSums()

}
\seealso{
\url{https://github.com/jrosell/jrrosell/blob/main/R/pipe.R}
}
