% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{theme_set_roboto_darkblue}
\alias{theme_set_roboto_darkblue}
\title{Sets a dark blue colored dark minimal theme using the Roboto font family}
\usage{
theme_set_roboto_darkblue(...)
}
\arguments{
\item{...}{Other parameters passed to theme_set}
}
\description{
Sets a dark blue colored dark minimal theme using the Roboto font family
}
\examples{
library(jrrosell)
library(ggplot2)
theme_set_roboto_darkblue()
ggplot(iris, aes(Species)) +
  geom_bar()

}
\keyword{graphing}
