% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{tokenize_text}
\alias{tokenize_text}
\title{Tokenize text}
\usage{
tokenize_text(text, sep = "\\\\s+")
}
\arguments{
\item{text}{A character vector or object that can be coerced to a character string. Represents the input text to be cleaned.}

\item{sep}{= "\\s+"}
}
\value{
A character vector
}
\description{
This function generates a character vector for a given text string
}
\examples{
# Example usage:
normalize_text("¡Hola! Esto es una prueba 123.") |> tokenize_text()

}
\keyword{text}
