% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{update_step}
\alias{update_step}
\title{Update recipe step values by id}
\usage{
update_step(object, target_id, ...)
}
\arguments{
\item{object}{A recipe or a workflow object with a recipe}
}
\description{
Update the vaules of a specific recipe step located by id
}
\examples{
recipes::recipe(spray ~ ., data = InsectSprays) |>
   recipes::step_ns(count, deg_free = hardhat::tune(), id="ns") |>
   update_step("ns", deg_free = 1)

}
\seealso{
https://github.com/jrosell/jrrosell/blob/main/R/tidymodels.R
}
