% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{workflow_boost_tree}
\alias{workflow_boost_tree}
\title{Create an xgboost tunable workflow for regression and classification}
\usage{
workflow_boost_tree(rec, engine = "xgboost", counts = TRUE, ...)
}
\arguments{
\item{rec}{prerocessing recipe to build the workflow}

\item{engine}{xgboost, lightgbm (xgboost by default)}

\item{counts}{Optional logic argument wether mtry use counts or not}

\item{...}{optional engine arguments}
}
\description{
Create an xgboost tunable workflow for regression and classification
}
\keyword{tidymodels}
