% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{workflow_elasticnet}
\alias{workflow_elasticnet}
\title{Create a tuneable glmnet worfklow for regression and classification}
\usage{
workflow_elasticnet(rec, engine = "glmnet", ...)
}
\arguments{
\item{rec}{prerocessing recipe to build the workflow}

\item{engine}{glmnet, spark, brulee (glmnet by default)}

\item{...}{Optional engine arguments}
}
\description{
Create a tuneable glmnet worfklow for regression and classification
}
\keyword{tidymodels}
