% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Write a list of tibbles to a xlsx file}
\usage{
write_xlsx(data, distfile, ...)
}
\arguments{
\item{data}{A named list of tibbles}

\item{distfile}{The name of the destination file.}

\item{...}{Other parameters to openxls::write.xlsx function}
}
\description{
It's useful for saving multiple data to a multiple sheets of a single Excel/Openoffice/libreoffice file.
}
\details{
The write_xlsx it's a wroapper for \code{openxls::write.xlsx}.
}
\examples{
l <- list("IRIS" = iris, "MTCATS" = mtcars, matrix(runif(1000), ncol = 5))
write_xlsx(l, "/tmp/writeList1.xlsx", colWidths = c(NA, "auto", "auto"))
file.remove("/tmp/writeList1.xlsx")

}
\seealso{
\url{https://github.com/jrosell/jrrosell/blob/main/R/write.R}
}
\keyword{write}
