% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_split_size}
\alias{calc_split_size}
\title{Calculate split size}
\source{
\url{https://stats.stackexchange.com/a/304996/7387}
}
\usage{
calc_split_size(
  std_err = NULL,
  confidence_interval = 0.95,
  margin_error = 0.02
)
}
\arguments{
\item{std_err}{The desired std_err numeric (default NULL)}

\item{confidence_interval}{(default 0.95)}

\item{margin_error}{(default 0.02)}
}
\description{
From binary classification problems, with the desired std_err it returns the
minimal assesment/validation set size.
}
\details{
The calc_validation_size function returns the minimal validation size for expected probabilities and desired error.
s
}
\examples{
calc_split_size()
calc_split_size(confidence_interval = 0.95, margin_error = 0.02)
calc_split_size(std_err = 0.02)

}
\keyword{tidymodels}
