% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{fit_results}
\alias{fit_results}
\title{Fit a workflow with specific parameters}
\usage{
fit_results(wf, resamples, param_info = NULL, grid = 10, fn = "tune_grid", ...)
}
\arguments{
\item{wf}{workflow}

\item{resamples}{rset}

\item{param_info}{for tune_* functions}

\item{grid}{for tune_* functions}

\item{fn}{the name of the function to run when tuning}

\item{...}{Optional engine arguments}
}
\description{
Fit a workflow with specific parameters
}
\examples{
library(tidymodels)
library(xgboost)
library(modeldata)
data(cells)
split <- cells |>
  mutate(across(where(is.character), as.factor)) |>
  sample_n(500) |>
  initial_split(strata = case)
train <- training(split)
resamples <- vfold_cv(train, v = 2, strata = case)
wf_spec <- train |>
  recipe(case ~ .) |>
  step_integer(all_nominal_predictors()) |>
  workflow(boost_tree(mode = "classification"))
res_spec <- wf_spec |> fit_results(resamples)
res_spec |> collect_metrics()
}
\keyword{tidymodels}
