% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{last_fit_metrics}
\alias{last_fit_metrics}
\title{Do the last fit and get the metrics}
\usage{
last_fit_metrics(res, split, metric)
}
\arguments{
\item{res}{Tune results}

\item{split}{The initial split object}

\item{metric}{What metric to use to select the best workflow}
}
\description{
Do the last fit and get the metrics
}
\examples{
library(tidymodels)
library(xgboost)
library(modeldata)
data(cells)
split <- cells |>
  mutate(across(where(is.character), as.factor)) |>
  sample_n(500) |>
  initial_split(strata = class)
train <- training(split)
folds <- vfold_cv(train, v = 2, strata = class)
wf <- train |>
  recipe(case ~ .) |>
  step_integer(all_nominal_predictors()) |>
  workflow_boost_tree()
res <- wf |>
  tune::tune_grid(
    resamples = folds,
    grid = 2,
    metrics = metric_set(roc_auc),
    control = tune::control_grid(save_workflow = TRUE, verbose = FALSE)
  )
res |> collect_metrics()
res |> last_fit_metrics(split, "roc_auc")
best <- res |> fit_best()
best |>
  augment(testing(split)) |>
  roc_auc(case, .pred_Test) |>
  pull(.estimate)

}
\keyword{tidymodels}
