% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{mapParallel}
\alias{mapParallel}
\title{Map parallel processing}
\usage{
mapParallel(x, fun_list)
}
\arguments{
\item{x}{An object}

\item{fun_list}{A list of functions to run in parallel over the object}
}
\description{
Map data over a function in one parallel core
}
\details{
The mapParallel function uses parallel
package to run functions to run in parallel over the object. It doesn't returning anything.
}
\examples{
data.frame(x = 2) |> mapParallel(list(function(x) print(x), function(x) print(x*2)))

}
\seealso{
\url{https://github.com/jrosell/jrrosell/blob/main/R/parallel.R}
}
