% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{package_github_version}
\alias{package_github_version}
\title{Github version of the package}
\usage{
package_github_version(x)
}
\arguments{
\item{x}{a single package}
}
\description{
Get the version from the DESCRIPTION file of the master branch in the github package
}
\seealso{
https://github.com/jrosell/jrrosell/blob/main/R/packages.R
}
\keyword{processing}
