% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot a variable}
\usage{
plot_variable(df, variable, ..., type = "numeric")
}
\arguments{
\item{df}{a data.frame}

\item{variable}{the variable to use.}

\item{...}{params passed to geom_*}

\item{type}{numeric (default) or nominal.}
}
\description{
It returns a bar or a histogram of the variable
}
\examples{
data.frame(a = c("x", "y", "y"), b = c("z", "z", "x")) |> plot_variable(a)
}
\keyword{wrangling}
