% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{prep_juice}
\alias{prep_juice}
\title{Prep, juice and glimpse a recipe or workflow}
\usage{
prep_juice(object)
}
\arguments{
\item{object}{A recipe or a workflow object with a recipe}
}
\description{
Stop started timer and parallel processing.
}
\details{
The stopParallel function uses tictoc and  parallel
packages. It returns the result of parallel::stopCluster(cl) method.
}
\examples{
recipes::recipe(spray ~ ., data = InsectSprays) |>
   prep_juice()
recipes::recipe(spray ~ ., data = InsectSprays)  |>
   workflows::workflow(parsnip::linear_reg()) |>
   prep_juice()

}
\seealso{
https://github.com/jrosell/jrrosell/blob/main/R/parallel.R
}
