% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_chr}
\alias{read_chr}
\title{Read character columns with clean names}
\usage{
read_chr(
  file,
  delim = ",",
  locale = NULL,
  ...,
  date_names = "en",
  date_format = "\%AD",
  time_format = "\%AT",
  decimal_mark = ".",
  grouping_mark = "",
  tz = "CET",
  encoding = "UTF-8",
  asciify = FALSE
)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either a single string or a raw vector).}

\item{delim}{Single character used to separate fields within a record.}

\item{locale}{The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.}

\item{...}{Other parameters to readr::read_delim.}

\item{date_names}{"en"}

\item{date_format}{"\%AD"}

\item{time_format}{"\%AT"}

\item{decimal_mark}{"."}

\item{grouping_mark}{""}

\item{tz}{"CET"}

\item{encoding}{"UTF-8"}

\item{asciify}{FALSE}
}
\description{
It's useful for reading the most common types of flat file data,
comma separated values and tab separated values.
}
\details{
The read_chr function works like \code{readr::read_delim}, except that
column sreturned would be characters and with clean names. It requires readr and janitor packages installed.
}
\examples{
read_chr(readr::readr_example("mtcars.csv"), delim = ",")

}
\keyword{read}
