% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{sum_missing}
\alias{sum_missing}
\title{Sum the missing values from a data.frame}
\usage{
sum_missing(...)
}
\arguments{
\item{...}{one or multiple data.frame}
}
\description{
Sum the missing values from a data.frame
}
\keyword{processing}
