% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{summarize_n_distinct}
\alias{summarize_n_distinct}
\title{Select constant columns from a data.frame}
\usage{
summarize_n_distinct(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Select constant columns from a data.frame
}
\examples{
summarize_n_distinct(data.frame(a = c(1, 2), b = c(2, 3)))
summarize_n_distinct(data.frame(a = c(1, 1), b = c(2, 3)))
}
\keyword{wrangling}
