% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{theme_roboto}
\alias{theme_roboto}
\title{Sets a minimal theme using the Roboto font family}
\usage{
theme_roboto(
  base_size = 11,
  strip_text_size = 12,
  strip_text_margin = 5,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  ...
)
}
\arguments{
\item{base_size}{= 11}

\item{strip_text_size}{= 12}

\item{strip_text_margin}{= 5}

\item{subtitle_size}{= 13}

\item{subtitle_margin}{= 10}

\item{plot_title_size}{= 16}

\item{plot_title_margin}{= 10}

\item{...}{Other parameters passed to theme_set}
}
\description{
It requires roboto fonts installed in your O.S. and run z
}
\keyword{graphing}
